:sparkles: Add test function decl test case
