--- Tests/RagnarokCoreTests/TestData/TestFunctionDecl.swift
@@ -1,31 +1,0 @@
-import Foundation
-
-public class TestFunctionDecl {
-    func noArgumentNoReturn() {
-        
-    }
-    func noArgumentHasReturnKeyword() -> Void {
-        
-    }
-    func oneArgumentNoReturn(argument: Int) {
-        
-    }
-    func oneArgumentHasReturnKeyword(argument: Int) -> Void {
-        
-    }
-    func twoArgumentNoReturn(argument1: Int, argument2: String) {
-        
-    }
-    func twoArgumentHasReturnKeyword(argument1: Int, argument2: String) -> Void {
-        
-    }
-    func noArgumentUsingThrowsKeyword() throws {
-        
-    }
-    func oneArgumentUsingThrowsKeyword(argument1: Int) throws {
-        
-    }
-    func multipleArgumentUsingThrowsKeyword(argument1: Int, argument2: String) throws {
-        
-    }
-}
