+++ Tests/RagnarokCoreTests/TestData/TestFunctionDeclHasReturnKeyword.swift
@@ -1,0 +1,20 @@
+//
+//  TestFunctionDeclHasReturnKeyword.swift
+//  RagnarokCoreTests
+//
+//  Created by Yudai.Hirose on 2019/01/13.
+//
+
+import Foundation
+
+public class TestFunctionDeclHasReturnKeyword {
+    func noArgumentHasReturnKeyword() -> Void {
+        
+    }
+    func oneArgumentHasReturnKeyword(argument: Int) -> Void {
+        
+    }
+    func twoArgumentHasReturnKeyword(argument1: Int, argument2: String) -> Void {
+        
+    }
+}
