+++ Tests/RagnarokCoreTests/TestData/TestFunctionDeclNoReturn.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public class TestFunctionDeclNoReturn {
+    func noArgumentNoReturn() {
+        
+    }
+    func oneArgumentNoReturn(argument: Int) {
+        
+    }
+    func twoArgumentNoReturn(argument1: Int, argument2: String) {
+        
+    }
+}
