+++ Tests/RagnarokCoreTests/TestData/TestFunctionDeclUsingThrows.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+public class TestFunctionDeclUsingThrows {
+    func noArgumentUsingThrowsKeyword() throws {
+        
+    }
+    func oneArgumentUsingThrowsKeyword(argument1: Int) throws {
+        
+    }
+    func multipleArgumentUsingThrowsKeyword(argument1: Int, argument2: String) throws {
+        
+    }
+}
